requirements.txt
google-generativeai>=0.8.0
python-dotenv==1.0.0
streamlit==1.29.0
SpeechRecognition==3.10.0
pyaudio==0.2.13